api/requirements.txt
appdirs==1.4.3
argon2-cffi==16.3.0
certifi==2017.4.17
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
coverage==4.5.1
Django==2.1.7
django-filter==2.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
msgpack-python==0.4.8
openapi-codec==1.3.1
packaging==16.8
Pillow==5.0.0
psycopg2-binary==2.7.7
pycparser==2.18
PyJWT==1.7.1
pyparsing==2.2.0
python-magic==0.4.15
pytz==2017.3
requests==2.18.1
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
